Revert "split up files so they can better be reused and tree-shaked"
This reverts commit 3ea22abcaac875479ce9ead6adbd0b59f1f22239.
